feat(PC-11584) Add LayoutExpiredLink
