+++ src/features/profile/pages/ChangeEmail/ChangeEmailExpiredLink.tsx
@@ -1,22 +1,14 @@
-import { t } from '@lingui/macro'
 import React from 'react'
-import styled from 'styled-components/native'
 
-import { GenericInfoPage } from 'ui/components/GenericInfoPage'
-import { SadFace } from 'ui/svg/icons/SadFace'
-import { ColorsEnum, Typo } from 'ui/theme'
+import { LayoutExpiredLink } from 'features/profile/pages/ChangeEmail/LayoutExpiredLink'
 
 export function ChangeEmailExpiredLink() {
-  return (
-    <GenericInfoPage title={t`Oups`} icon={SadFace}>
-      <StyledBody>{t`Le lien est expiré !`}</StyledBody>
-      <StyledBody>{t`Clique sur « Renvoyer l’e-mail » pour recevoir un nouveau lien.`}</StyledBody>
-    </GenericInfoPage>
-  )
-}
+  const changeEmailExpiredLink = () => {
+    // TODO (LucasBeneston): try / catch api call like api.postnativev1requestPasswordReset({ email })
+  }
 
-const StyledBody = styled(Typo.Body).attrs({
-  color: ColorsEnum.WHITE,
-})({
-  textAlign: 'center',
-})
+  // TODO (LucasBeneston): use isFetching from useQuery()
+  const isFetching = true
+
+  return <LayoutExpiredLink resetQuery={changeEmailExpiredLink} isFetching={isFetching} />
+}
